+++ src/main/java/com/example/assignment2/entity/Customer.java
@@ -7,7 +7,7 @@ import javax.persistence.*;
 public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "customer_Id")
     private String customerID;
@@ -25,15 +25,15 @@ public class Customer {
     public Customer() {
     }
 
-    public Customer(Integer id) {
+    public Customer(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
