+++ src/main/java/com/example/assignment2/entity/Invoice.java
@@ -7,7 +7,7 @@ import javax.persistence.*;
 public class Invoice  {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name ="invoiceID")
     private String invoiceID;
@@ -21,7 +21,7 @@ public class Invoice  {
     public Invoice() {
     }
 
-    public Invoice(Integer id, String invoiceID, String driverID, String customerID, String totalCharge) {
+    public Invoice(Long id, String invoiceID, String driverID, String customerID, String totalCharge) {
         this.id = id;
         this.invoiceID = invoiceID;
         this.driverID = driverID;
@@ -29,11 +29,11 @@ public class Invoice  {
         this.totalCharge = totalCharge;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
