+++ src/main/java/com/example/assignment2/repository/CarRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface CarRepository extends JpaRepository<Car, Integer> {
+public interface CarRepository extends JpaRepository<Car, Long> {
 
 
-    Optional<Car> findById(Integer id);
+    Optional<Car> findById(Long id);
 }
