+++ src/main/java/com/example/assignment2/repository/DriverRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 import java.util.Optional;
-public interface DriverRepository extends JpaRepository<Driver, Integer> {
-    Optional<Driver> findById(Integer id);
+public interface DriverRepository extends JpaRepository<Driver, Long> {
+    Optional<Driver> findById(Long id);
 
 
 }
